split TerminalFeedContext what and why into PerceptionManager
